requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
langserve
fastapi 
uvicorn
sse_starlette